Refactor Item Type Test Cases
