+++ Assets/Tests/EditMode/ItemType.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5747e0907f084484baba8848aa63fee6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
