+++ Assets/Tests/EditMode/ItemType/TestItemTypeAddType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e35125b8a8bbdc14bae936213d2130db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
