+++ Assets/Tests/EditMode/ItemType/TestItemTypeInitialization.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ecd292fc0e6ca8c8a80c03afbd24203a
+guid: ec5cab03b662c8540be3297ea35888df
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
