+++ Assets/Tests/EditMode/ItemType/TestItemTypeRemoveType.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3d18e4cf95f08674ba72ae419dbf8842
+guid: b09ac81e6270e5a4fb07f5813f06b1c6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
