+++ Assets/Tests/EditMode/ItemType/TestItemTypeSetType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d2987dd76e6c434bbc5a7eaf7ecf929
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
